ex00 checked and ready to deliver

+++ cpp00/ex00/megaphone.cpp
@@ -7,7 +7,7 @@ int main(int ac, char **av)
 	if (ac == 1)
 		std::cout << "* LOUD AND UNBEARABLE FEEDBACK NOISE *" << std::endl;
 	else
-		for (size_t i = 1; i < ac; i++)
+		for (int i = 1; i < ac; i++)
 			for (size_t j = 0; j <  strlen(av[i]); j++)
 				std::cout << (char)toupper(av[i][j]);
